works.
satisfies the following minimum requirements:
The library of available files that contain MP3 encoded music must appear on the screen in a manner such that the user can select which song to start playing. After that song completes, the next song in the library should start playing after a short silence period. This should continue until all of the songs in the library after the first selected song are played.
The user should be able to pause and restart the playing of music via the pressing of a certain button on the user interface.
The user should be able to jump forward to the next song, or back to the previous song, via buttons on the user interface.
The user should be able to make “play lists.” Each play list would contain the names of the files that the user wishes to group together into a set of songs to be played. As with the library, after the first song completes, the next song on the list should start playing.
Play lists should have the option of either being circular (that is, once the last song on the list is played, the first one is then played. This “looping” would allow a play list to be played at a party all night long) or linear (stop after the last song on the list. This feature should be a “ON/OFF” toggle, which is achieved via a particular button on the user interface.
A helpful menu system should be included.